requirements.txt
ynab-sdk==0.5.0
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0
python-dotenv==1.0.0
datetime
flask==2.3.3 